Small WindowingAPI fixes 07/26/2021 | 21w30a1 | 0.7.62

+++ TRAP/src/Events/KeyEvent.cpp
@@ -210,14 +210,14 @@ std::string TRAP::Events::KeyEvent::NonPrintableKeyToString(const Input::Key key
 //-------------------------------------------------------------------------------------------------------------------//
 //-------------------------------------------------------------------------------------------------------------------//
 
-TRAP::Events::KeyPressEvent::KeyPressEvent(const Input::Key key, const uint16_t repeatCount, TRAP::Window* window)
+TRAP::Events::KeyPressEvent::KeyPressEvent(const Input::Key key, const uint32_t repeatCount, TRAP::Window* window)
 	: KeyEvent(key), m_repeatCount(repeatCount), m_window(window)
 {
 }
 
 //-------------------------------------------------------------------------------------------------------------------//
 
-uint16_t TRAP::Events::KeyPressEvent::GetRepeatCount() const
+uint32_t TRAP::Events::KeyPressEvent::GetRepeatCount() const
 {
 	return m_repeatCount;
 }
